requirements.txt
aiosqlite==0.17.0
alembic==1.8.0
anyio==3.6.1
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
decorator==5.1.1
executing==0.8.3
fastapi==0.78.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
ipython==8.4.0
jedi==0.18.1
Mako==1.2.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
pytest-asyncio==0.18.3
pytest-dotenv==0.5.2
python-dotenv==0.20.0
requests==2.28.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
sqlalchemy2-stubs==0.0.2a19
sqlmodel==0.0.6
stack-data==0.3.0
starlette==0.19.1
tomli==2.0.1
traitlets==5.3.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.18.2
wcwidth==0.2.5
